requirements.txt
alembic==0.9.9
bcrypt==3.1.2
cffi==1.9.1
click==6.6
coverage==4.2
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.6.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.6
MarkupSafe==1.0
pkg-resources==0.0.0
psycopg2==2.6.2
pycparser==2.17
PyJWT==1.4.2
python-dotenv==0.8.2
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
